requirements.txt
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.2
docling>=1.0.0
faiss-cpu>=1.7.4
streamlit>=1.28.0
langchain>=0.1.0
chromadb>=0.4.0
llama-cpp-python>=0.2.0
numpy>=1.24.0
pandas>=2.0.0
pypdf2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0